+++ R/libr-package.R
@@ -1,0 +1,5 @@
+## usethis namespace: start
+#' @importFrom Rcpp sourceCpp
+#' @useDynLib libr
+## usethis namespace: end
+NULL