+++ tests/testthat/test-datastep.R
@@ -18,7 +18,7 @@ test_that("datastep() function works as expected with mtcars.", {
 
   })
 
-
+  
   expect_equal("mpgcat" %in% names(d1), TRUE)
 
 })
@@ -265,35 +265,7 @@ test_that("datastep() retain class check works as expected", {
 
 })
 
-test_that("datastep() performance is good", {
 
-  if (DEV) {
-
-    libname(dat, file.path(base_path, "SDTM"), "sas7bdat")
-
-    tm <- Sys.time()
-
-    res <- datastep(dat$lb, retain = list(rnum = 0),
-                    by = c("USUBJID"),
-                    {
-      if (first.)
-        subjstart <- TRUE
-      else
-        subjstart <- FALSE
-
-      rnum <- rnum + 1
-
-    })
-
-    tmdiff <- Sys.time() - tm
-    tmdiff
-
-    expect_equal(tmdiff < 16, TRUE)
-
-  } else
-    expect_equal(TRUE, TRUE)
-
-})
 
 
 test_that("Rename works as expected", {
@@ -314,64 +286,6 @@ test_that("Rename works as expected", {
 
 })
 
-
-test_that("datastep() with group_by performance is good", {
-  
-  if (DEV) {
-    
-    library(dplyr)
-    
-    scs <- specs(PE = import_spec(PESTAT = "character"))
-    
-    libname(dat, file.path(base_path, "SDTM"), "csv", import_specs = scs)
-    
-    tm <- Sys.time()
-    
-    prep <- dat$DM %>% 
-      left_join(dat$VS, by = c("USUBJID" = "USUBJID")) %>% 
-      select(USUBJID, VSTESTCD, VISIT, VISITNUM, VSSTRESN, ARM, VSBLFL) %>% 
-      filter(VSTESTCD %in% c("PULSE", "RESP", "TEMP", "DIABP", "SYSBP"), 
-             !(VISIT == "SCREENING" & VSBLFL != "Y")) %>% 
-      arrange(USUBJID, VSTESTCD, VISITNUM) %>% 
-      group_by(USUBJID, VSTESTCD) %>%
-      #datastep(by = c("USUBJID", "VSTESTCD"), retain = list(BSTRESN = 0), {
-      datastep(retain = list(BSTRESN = 0), {
-
-        # Combine treatment groups
-        # And distingish baseline time points
-        if (ARM == "ARM A") {
-
-          if (VSBLFL %eq% "Y") {
-            GRP <- "A_BASE"
-          } else {
-            GRP <- "A_TRT"
-          }
-
-        } else {
-
-          if (VSBLFL %eq% "Y") {
-            GRP <- "O_BASE"
-          } else {
-            GRP <- "O_TRT"
-          }
-
-        }
-
-        # Populate baseline value
-        if (first.)
-          BSTRESN = VSSTRESN
-
-      })
-    
-    tmdiff <- Sys.time() - tm 
-    tmdiff
-    
-    expect_equal(tmdiff < 20, TRUE)
-    
-  } else 
-    expect_equal(TRUE, TRUE)
-  
-})